User/OpenIDConnect: Fix authentication
See:
- https://mantis.ilias.de/view.php?id=34951
- https://mantis.ilias.de/view.php?id=35066

+++ Services/OpenIdConnect/classes/class.ilOpenIdConnectUserSync.php
@@ -50,7 +50,7 @@ class ilOpenIdConnectUserSync
     public function setInternalAccount(string $int_account): void
     {
         $this->int_account = $int_account;
-        $this->usr_id = ilObjUser::_lookupId($this->int_account);
+        $this->usr_id = (int) ilObjUser::_lookupId($this->int_account);
     }
 
     public function getUserId(): int
